refactor: extract logging classes
* RobolectricDebugTree
* ProductionCrashReportingTree
* FragmentLifecycleLogger
